todoapp2/requirements.txt
asgiref==3.6.0
Django==3.2.8
django-crispy-forms==2.0
modeltranslation==0.25
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
